requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
# python-jose[cryptography] # Deprecated
passlib[bcrypt]>=1.7.4
pinecone>=5.3.0
pypdf>=4.3.1
httpx>=0.27.0
python-multipart>=0.0.12
pydantic>=2.9.0
streamlit>=1.38.0
requests>=2.32.3
PyJWT>=2.9.0
cryptography>=43.0.0
python-dotenv>=1.0.1
